requirements.txt
aiohttp>=3.9.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
certifi==2025.8.3
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.12.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai>=1.0.0
packaging==25.0
pillow==12.0.0
pyasn1==0.6.1
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pytesseract==0.3.13
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rsa==4.9.1
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1